Add validation and more checks for existing intermediate results

+++ src/main.py
@@ -11,42 +11,48 @@ from src.config.config import Config
 from src.diagnostic_ions.detection import DiagnosticIonDetector
 from src.diagnostic_ions.summary import (
     get_detected_modifications_with_combinations,
-    plot_detected_ion_combinations,
-    plot_detected_ions,
-)
+    plot_detected_ion_combinations, plot_detected_ions)
 from src.diagnostic_ions.utils import get_mod_combination_str
 from src.mzml_processing.extraction import ScanWindowExtractor
 from src.mzml_processing.utils import get_diann_compatible_mzml_output_file
 from src.split_processing.result_aggregation import ResultAggregation
 from src.split_processing.scan_window_splitting import ScanWindowSplitting
-from src.split_processing.spectral_library_splitting import split_library_by_mods
-from src.utils import (
-    make_dia_nn_additional_param_commands,
-    make_dia_nn_var_mod_commands,
-)
+from src.split_processing.spectral_library_splitting import \
+    split_library_by_mods
+from src.utils import (make_dia_nn_additional_param_commands,
+                       make_dia_nn_var_mod_commands)
+from src.validation import (validate_filepath_exists,
+                            validate_modification_combinations,
+                            validate_modifications,
+                            validate_number_lower_higher_energy_windows,
+                            validate_spectral_library_files_by_mod,
+                            validate_spectral_library_not_empty)
 
 
 def main(config: Config, logger: logging.Logger):
     result_path = Path(config.result_dir)
 
+    validate_filepath_exists(Path(config.mzml_file))
+    validate_filepath_exists(Path(config.dia_nn_path))
+
     logger.info("Loading spectra...")
 
     exp = MSExperiment()
-    MzMLFile().load(str(config.mzml_file), exp)
+    MzMLFile().load(config.mzml_file, exp)
 
     extractor = ScanWindowExtractor(
         config.lower_collision_energy, config.higher_collision_energy
     )
 
-    # TODO: add some validation (including whether all paths exist etc)
-    # and create result dir if it doesnt exist
     ms1_and_higher_energy_windows = extractor.extract_ms1_and_higher_energy_windows(exp)
     higher_energy_windows = extractor.extract_higher_energy_windows(
         ms1_and_higher_energy_windows
     )
     ms1_windows = extractor.extract_ms1_windows(ms1_and_higher_energy_windows)
-
     ms1_and_lower_energy_windows = extractor.extract_ms1_and_lower_energy_windows(exp)
+    validate_number_lower_higher_energy_windows(
+        ms1_and_lower_energy_windows, ms1_and_higher_energy_windows
+    )
 
     # ---------------------- Diagnostic ion detection -------------------
 
@@ -85,7 +91,9 @@ def main(config: Config, logger: logging.Logger):
     if config.modifications_to_search != []:
         modifications_to_search = config.modifications_to_search
         modification_combinations = config.modification_combinations
-        # TODO: add validation that config mod combinations are not single mods
+
+        validate_modifications(modifications_to_search)
+        validate_modification_combinations(modification_combinations)
     else:
         modifications_to_search, modification_combinations = (
             get_detected_modifications_with_combinations(
@@ -96,7 +104,6 @@ def main(config: Config, logger: logging.Logger):
             )
         )
 
-    # TODO: test it
     logger.info(
         "Keeping ions for mods %s.",
         set(modifications_to_search).union(set.union(*modification_combinations)),
@@ -108,7 +115,8 @@ def main(config: Config, logger: logging.Logger):
     ]
 
     logger.info(
-        "Considering single mods %s and combinations %s for window splitting and spectral library preparation.",
+        "Considering single mods %s and combinations %s "
+        "for window splitting and spectral library preparation.",
         modifications_to_search,
         [
             get_mod_combination_str(mod_combination)
@@ -117,6 +125,8 @@ def main(config: Config, logger: logging.Logger):
     )
 
     if len(config.modifications_additional) > 0:
+        validate_modifications(config.modifications_additional)
+
         if config.library_free:
             library_log_text = "and added to spectral library prediction"
         else:
@@ -136,16 +146,16 @@ def main(config: Config, logger: logging.Logger):
     # ---------------------- Spectral library preparation -------------------
 
     if config.library_free:
-        # TODO: validate that the path is there
+        logger.info(
+            "Library-free mode was selected. Spectral libraries are predicted by DIA-NN."
+        )
+
         database_path = config.database_for_library_prediction
+        validate_filepath_exists(Path(database_path))
         mods_for_lib_creation = np.concatenate(
             [modifications_to_search, modification_combinations, ["unmodified"]]
         )
         spectral_library_files_by_mod = {}
-
-        logger.info(
-            "Library-free mode was selected. Spectral libraries are predicted by DIA-NN."
-        )
 
         for mods in mods_for_lib_creation:
             mod_combination_str = get_mod_combination_str(mods)
@@ -154,71 +164,80 @@ def main(config: Config, logger: logging.Logger):
             )
             predicted_library_path = Path(f"{library_path_for_diann}.predicted.speclib")
 
-            if not predicted_library_path.exists():
-                var_mod_commands_for_lib = make_dia_nn_var_mod_commands(
-                    mods, config.modifications_additional
-                )
+            if predicted_library_path.exists():
+                logger.info("Spectral library for mod(s) %s already exists.", mod_combination_str)
+                continue
 
-                additional_param_commands_lib = make_dia_nn_additional_param_commands(
-                    config.dia_nn_library_params
-                )
+            var_mod_commands_for_lib = make_dia_nn_var_mod_commands(
+                mods, config.modifications_additional
+            )
 
-                dia_nn_library_command_for_mod = (
-                    [
-                        f"{config.dia_nn_path}",
-                        "--gen-spec-lib",
-                        "--fasta-search",
-                        "--predictor",
-                        "--strip-unknown-mods",
-                        # DIA-NN does automatic extension removal for the library path,
-                        # i.e. takes the path before the last '.', so if there is a '.'
-                        # in the path to the result directory, the resulting library
-                        # path is messed up. That is the reason
-                        # for this unnecessary .predicted here.
-                        f"--out-lib {library_path_for_diann}.predicted",
-                        f"--fasta {database_path}",
-                    ]
-                    + additional_param_commands_lib
-                    + var_mod_commands_for_lib
-                )
+            additional_param_commands_lib = make_dia_nn_additional_param_commands(
+                config.dia_nn_library_params
+            )
 
-                logger.info(
-                    "Running DIA-NN to create spectral library for mod(s) %s  with command %s.",
-                    mod_combination_str,
+            dia_nn_library_command_for_mod = (
+                [
+                    f"{config.dia_nn_path}",
+                    "--gen-spec-lib",
+                    "--fasta-search",
+                    "--predictor",
+                    "--strip-unknown-mods",
+                    # DIA-NN does automatic extension removal for the library path,
+                    # i.e. takes the path before the last '.', so if there is a '.'
+                    # in the path to the result directory, the resulting library
+                    # path is messed up. That is the reason
+                    # for this unnecessary .predicted here.
+                    f"--out-lib {library_path_for_diann}.predicted",
+                    f"--fasta {database_path}",
+                ]
+                + additional_param_commands_lib
+                + var_mod_commands_for_lib
+            )
+
+            logger.info(
+                "Running DIA-NN to create spectral library for mod(s) %s  with command %s.",
+                mod_combination_str,
+                dia_nn_library_command_for_mod,
+            )
+            logger.info(
+                "The logs of the run will be found in the respective DIA-NN logfile. "
+                "Starting run..."
+            )
+            try:
+                subprocess.run(
                     dia_nn_library_command_for_mod,
+                    check=True,
                 )
-                logger.info(
-                    "The logs of the run will be found in the respective DIA-NN logfile. Starting run..."
-                )
-                try:
-                    subprocess.run(
-                        dia_nn_library_command_for_mod,
-                        check=True,
-                    )
-                except subprocess.CalledProcessError as e:
-                    logger.exception(
-                        "DIA-NN spectra library creation for mod(s) %s crashed.",
-                        mod_combination_str,
-                    )
-                    raise e
-
-                logger.info(
-                    "Spectral library for mod(s) %s has been created.",
+            except subprocess.CalledProcessError as e:
+                logger.exception(
+                    "DIA-NN spectra library creation for mod(s) %s crashed.",
                     mod_combination_str,
                 )
+                raise e
+
+            logger.info(
+                "Spectral library for mod(s) %s has been created.",
+                mod_combination_str,
+            )
 
             spectral_library_files_by_mod[mods] = predicted_library_path
         logger.info("Prediction of all required spectral libraries has finished.")
-
     else:
         if config.spectral_library_files_by_mod:
-            # TODO: add validation that the keys match the mods (and combinations) to search and contains one "unmodified"
             logger.info(
                 "A spectral library per split was provided. Using those libraries for search."
+            )
+            validate_spectral_library_files_by_mod(
+                config.spectral_library_files_by_mod,
+                modifications_to_search,
+                modification_combinations,
             )
             spectral_library_files_by_mod = config.spectral_library_files_by_mod
         else:
             logger.info("Splitting provided spectral library...")
+            validate_filepath_exists(Path(config.spectral_library_for_filtering_path))
+
             library = pd.read_csv(
                 config.spectral_library_for_filtering_path, delimiter="\t"
             )
@@ -231,8 +250,8 @@ def main(config: Config, logger: logging.Logger):
             )
             spectral_library_files_by_mod = {}
 
-            # TODO: add validation/check if there are empty libraries
             for mods, library_df in spectral_library_df_by_mod.items():
+                validate_spectral_library_not_empty(library_df, get_mod_combination_str(mods))
                 library_path = (
                     result_path
                     / f"spectral_library_{get_mod_combination_str(mods)}.tsv"
@@ -283,11 +302,8 @@ def main(config: Config, logger: logging.Logger):
 
     file_paths_by_mods = {}
 
-    # TODO: skip run if result already exists
     for mods, windows_for_mod in windows_by_mods.items():
         mod_combination_str = get_mod_combination_str(mods)
-        logger.info("Preparing search for modification(s) %s...", mod_combination_str)
-
         dia_nn_report_path = result_path / f"report_{mod_combination_str}.tsv"
         spectrum_id_mapping_path = (
             result_path / f"spectrum_id_mapping_{mod_combination_str}.csv"
@@ -299,25 +315,34 @@ def main(config: Config, logger: logging.Logger):
             "spectrum_id_mapping_path": spectrum_id_mapping_path,
         }
 
-        logger.info(
-            "Renaming spectrum IDs to satisfy condition "
-            "of incremental IDs without missing numbers for DIA-NN."
-        )
-        windows_for_mod, spectrum_id_mapping = extractor.rename_spectrum_ids(
-            windows_for_mod, return_id_mapping=True
-        )
+        logger.info("Preparing search for modification(s) %s...", mod_combination_str)
 
-        output_file = get_diann_compatible_mzml_output_file()
-        output_file.store(str(mzml_path), windows_for_mod)
-        logger.info("Saved windows to search (MS1 and lower-energy) in %s.", mzml_path)
+        if mzml_path.exists() and spectrum_id_mapping_path.exists():
+            logger.info("MzML with windows to search and ID mapping file already exist.")
+        else:
+            logger.info(
+                "Renaming spectrum IDs to satisfy condition "
+                "of incremental IDs without missing numbers for DIA-NN."
+            )
+            windows_for_mod, spectrum_id_mapping = extractor.rename_spectrum_ids(
+                windows_for_mod, return_id_mapping=True
+            )
 
-        spectrum_id_mapping.to_csv(spectrum_id_mapping_path, index=False)
-        logger.info(
-            "Saved mapping from original to renamed spectrum IDs in %s.",
-            spectrum_id_mapping_path,
-        )
+            output_file = get_diann_compatible_mzml_output_file()
+            output_file.store(str(mzml_path), windows_for_mod)
+            logger.info("Saved windows to search (MS1 and lower-energy) in %s.", mzml_path)
+
+            spectrum_id_mapping.to_csv(spectrum_id_mapping_path, index=False)
+            logger.info(
+                "Saved mapping from original to renamed spectrum IDs in %s.",
+                spectrum_id_mapping_path,
+            )
 
         spectral_library_file_for_mod = spectral_library_files_by_mod[mods]
+
+        if dia_nn_report_path.exists():
+            logger.info("DIA-NN search report already exists.")
+            continue
 
         var_mod_commands_for_mods = make_dia_nn_var_mod_commands(
             mods, config.modifications_additional
@@ -367,6 +392,8 @@ def main(config: Config, logger: logging.Logger):
 
     # ---------------------- Aggregation -------------------
 
+    logger.info("Starting aggregation of DIA-NN reports...")
+
     aggregation = ResultAggregation(config.fdr_threshold, config.normalize_cscores)
     report_all_targets_with_decoys, report_fdr_filtered = aggregation.aggregate_results(
         result_path, file_paths_by_mods
@@ -396,11 +423,11 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     config = Config.from_path(Path(args.config_path))
+    result_path = Path(config.result_dir)
+    result_path.mkdir(exist_ok=True)
 
     logger = logging.getLogger(__name__)
-    logging.basicConfig(
-        filename=Path(config.result_dir) / "log.txt", level=logging.INFO
-    )
+    logging.basicConfig(filename=result_path / "log.txt", level=logging.INFO)
 
     try:
         main(config, logger)
